backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
slowapi==0.1.9
google-genai==0.2.2
python-docx==1.1.2
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-settings==2.6.1
